Update InputfieldRepeater.module
I find
Repeater-Name #1
Repeater-Name #2
Repeater-Name #n
is too generic and not enough when the wrappers are collapsed
i.e. for sorting purposes

So it looks if there is a title or a headline defined and
prepends it to the generic label.

Maybe you can think of yet another field besides title and headline.