Circular reference mitigation
Adding self check to prevent circular reference on page save leading to OOM. This will occur in cases where a Page contains a Page select input, with $this selected. While likely a mistake or the result of poor template design, it will white-screen any page save containing the field as it attempts to uncache. If such a reference should instead be illegal, throwing an error at this point could be the way to go?