Added enclosure
Hi Ryan,

This update add the possibility to add an enclosure to the RSS Markup generation. For the enclosure an image or a file field can be used. If the field for the enclosure is an image, the image can be scaled to preferred height & width.

If the checkbox Bounding box is checked the images will scale until one side reaches the goal dimension. Specified by width and height, useful for logo's.

Other fixes:
- changed uft to UTF in the xml header - W3c recomendation
- changed the dates to date(DATE_RSS) - W3c recomendation
- updated the version with 1

Thanks,
Martijn