Making <guid> the same as link, as link
Took quite a while why the module was not outputting items on the live server. **if(!$page->viewable()) continue;** in the renderFeed methode prevents this (pages had no template file). 

It's rather an odd one here, because if the find query returns Pages, I would expect those pages in the feed.