Selecting pages with no title
It is possible to use this _fieldtype_ for pages with no title (for example for selecting _users_). In such case, unless the fieldtype is configured to use something else than title, it displays empty items.
I've added a check for empty label, it falls back to _$page->id_. That, if I'm no mistaken, makes it somewhat coherent with _List of fields to display in the admin Page List_ field in templates' settings. Perhaps it would make more sense to add a _title_ to user pages (instead of _name_, maybe?) or replace the label with something akin to _undefined string_.