Fixed issue in SessionLoginThrottle.module
Reset the number of login attempts for a user if more than $requireSeconds have elapsed since his last login attempt. 
This fixes an issue than occurs if only one user logins to the  system during a certain period of time: in this case, the number of attempts for this user was always increased regardless of the time interval since the last login attempt. If the user entered a wrong password once, he was forced to wait the max time before retrying, even if his previous connection was several hours ago!