added support for 'forcenew' and 'suffix'
for pageimages it is direct into size, as helpers for third party
modules there are two functions that assist in adding a suffix and help
by finding a variation with a specified suffix.

@Ryan: there are some comments in the code in isVariation() regarding
regexp. I cannot solve it. Please can you do!

---------------------

here is simple testcode for the new methods:

https://gist.github.com/horst-n/ef70453e1007c172a162