Add support for same-line “block comments”
I’m sorry if this is silly, but it felt intuitive that one would be able to do this, and the additional code seems negligible. The use case is that one may want to include some HTML on the same line as a translatable string that isn’t supposed to show up on the translation page. Such as:

<p><?php echo __('ab'); /* As in “Starting at 90€” */ ?> <em><?php echo $page->price; ?></em></p>

It sometimes feels cleaner than storing the translated string in a one-off variable on a separate line.