clarify how _init.php is included in beginner profile
This really confused me for a while... based on the docs about the beginner profile I assumed there was some magic functionality of PW that always included an `_init.php` file (which seemed very un-PW-like to have such magic for the templates, so I'm glad that is not the case!), so I thought it might help other beginners to clarify the comments around this. Note that the addition to the comments is copy-pasted exactly from `/site-default/templates/_init.php`.

I could understand that maybe you want to keep this file sparse and simple for beginners -- in that case, I'd suggest at least clarifying how this works on the docs page: http://processwire.com/docs/tutorials/default-site-profile/page2

Thanks!