use cache folder for unvalidated tempfiles
@ryancramerdesign: regarding conversations with @teppokoivula, we should create the unvalidated files in the cache folder, not in (maybe thousands of) assets/files/ subfolders. If a site may proceed (for what ever reason) many of those unvalidated imagefiles, there is no chance to have them removed automatically on image-delete. 

Also with thousands of pages it is nearly impossible to clean those directories, or at least, it isn't worth the hassle. 

We can use the cache folder because we can relay of its existence and that it is writeable. We need not use subfolders (no use of mkdir etc). To have unique filenames across pages, we add the $page->id, and to keep imagesizers ability to detect filetypes upon filextension we add it again after the page->id. This way we would find orphaned unvalidated files in one central folder.

Thanks @teppokoivula!