Fixes sort for subvalues
There seemed to be a bug when trying to sort by a subvalue of a FieldtypePage, e.g., $pages->find("template=product, sort=brand.title"). I was getting an error "Unknown column '_sort_page_brand_title.title' in 'order clause'". The change fixes this, but I am not absolutely certain that it covers all bases, so someone with better knowledge about the table structure should have a look...