Proposal to implement ImageInspector
an centralized, independent, early invoked Image Inspector for
ImageSizer and the ImageSizerEngines.

This is useful for the ImageSizer and the ImageSizerEngines. PW now can
detect all sorts of Imageformats and subformats. If we also implement a
mandatory static feature list into ImageEngines, we can detect / select
the ImageSizerEngine directly, without possibly load one that are not
supports a given subformat. What (before ImageInspector) leads into
(re)loading the image into the next Engine.

@ryancramerdesign : I have implemented it to the state that the images
gets inspected early in the constructor of the ImageSizer and its result
is available in ImageSizer via the new method "". It is also available
in the Imagesizer Engines too.

Next step would be to add a feature list to the engines, containing what
type of subformats they support:
- gif, gif-trans, gif-anim, gif-anim-trans,
- png24, png24-trans, png8, png8-trans
- jpg

This adds a better performance for direct selection of the best engine
and it frees the engines of doing its own detection, as we provide all
needed information already!